Move some hooks usage to MessageEdit component and filter drop down list
