+++ ui/src/components/MessageEdit.tsx
@@ -1,20 +1,32 @@
 import React from 'react'
 import { Form, Input, Dropdown, Button } from 'semantic-ui-react';
-import { Text } from '@daml/types';
+import { Party } from '@daml/types';
 import { User } from '@daml2ts/create-daml-app/lib/create-daml-app-0.1.0/User';
+import { useParty, useExerciseByKey } from '@daml/react';
 
 type Props = {
-  friends: User[];
-  sendMessage: (content: Text, receiver: string) => Promise<boolean>;
+  users: Party[];
 }
 
 /**
  * React component to edit a message to send to a friend.
  */
-const MessageEdit: React.FC<Props> = ({friends, sendMessage}) => {
-  const [content, setContent] = React.useState('');
+const MessageEdit: React.FC<Props> = ({users}) => {
+  const sender = useParty();
   const [receiver, setReceiver] = React.useState('');
+  const [content, setContent] = React.useState('');
+  const [exerciseSendMessage] = useExerciseByKey(User.SendMessage);
   const [isSubmitting, setIsSubmitting] = React.useState(false);
+
+  const sendMessage = async (content: string, receiver: string): Promise<boolean> => {
+    try {
+      await exerciseSendMessage(receiver, {sender, content});
+      return true;
+    } catch (error) {
+      alert("Error sending message:\n" + JSON.stringify(error));
+      return false;
+    }
+  }
 
   const submitMessage = async (event?: React.FormEvent) => {
     if (event) {
@@ -24,19 +36,19 @@ const MessageEdit: React.FC<Props> = ({friends, sendMessage}) => {
     const success = await sendMessage(content, receiver);
     setIsSubmitting(false);
     if (success) {
-      setContent('');
       setReceiver('');
+      setContent('');
     }
   }
 
-  const friendOptions =
-    [...friends]
-    .map((friend) => friend.username)
+  // Sorted array of friends of the current user
+  const friends =
+    users
+    .filter(user => user !== sender)
     .sort((x, y) => x.localeCompare(y))
-    .map((friend) => (
-      { key: friend,
-        text: friend,
-        value: friend }));
+
+  // Options for dropdown menu
+  const friendOptions = friends.map((f) => ({ key: f, text: f, value: f }));
 
   return (
     <Form onSubmit={submitMessage}>
